-delete, insert, update 1차 리펙토링 완료
